Renamed StateMachine to StatefulEntity
