+++ CHANGELOG.md
@@ -27,7 +27,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Provided default no-op `TransactionContext` (#1461)
 * Refactored query capabilities for `Asset` (#1459)
 * Refactored query capabilities for `ContractDefinition` (#1458)
-* Refactored StateMachine and in-memory persistence (#1511)
+* Refactored state machine and in-memory persistence (#1511)
 
 #### Removed
